+++ odoo/local-src/connector_esb/models/backend.py
@@ -17,6 +17,7 @@ class ESBBackend(models.Model):
     sftp_host = fields.Char(string='SFTP Host', compute='_compute_from_env')
     sftp_port = fields.Integer(string='SFTP Port', compute='_compute_from_env')
     sftp_user = fields.Char(string='SFTP User', compute='_compute_from_env')
+    sftp_path = fields.Char(string='SFTP Path', compute='_compute_from_env')
     ws_url = fields.Char(
         string='Web Service Base URL',
         compute='_compute_from_env'
@@ -47,6 +48,7 @@ class ESBBackend(models.Model):
             record.sftp_host = os.getenv('ODOO_ESB_SFTP_HOST', '')
             record.sftp_port = int(os.getenv('ODOO_ESB_SFTP_PORT', 22))
             record.sftp_user = os.getenv('ODOO_ESB_SFTP_USER', '')
+            record.sftp_path = os.getenv('ODOO_ESB_SFTP_PATH', '')
             record.ws_user = os.getenv('ODOO_ESB_WS_USER', '')
             record.ws_pwd = os.getenv('ODOO_ESB_WS_PWD', '')
             record.ws_url = os.getenv('ODOO_ESB_WS_BASE_URL', '')