+++ HISTORY.rst
@@ -14,6 +14,7 @@ Unreleased
 
 * ALCN-1087: Disable Camptocamp platform specifics by default
   plus add an opt-in with a new environment variable C2C_PLATFORM
+* Generate invoices automatically every 10 days
 
 **Documentation**